fix: firmware compiler warnings

+++ pipit-firmware/Pipit.h
@@ -78,8 +78,8 @@ private:
   const uint8_t disconnect_readings_threshold = 3;
   uint8_t num_disconnect_readings = disconnect_readings_threshold;
 
-  const uint8_t connection_check_delay_short = 3000;
-  const uint8_t connection_check_delay_long = 6000;
+  const uint16_t connection_check_delay_short = 3000;
+  const uint16_t connection_check_delay_long = 6000;
 
 };
